fetch data and build log if needed
